+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+- Replace nokogiri with rexml + sanitize for html and xml handlers #26
+
 ## 3.0.0 - 2019-05-14
 ### Added
 - Add application configuration file containing version number